Add Virtual input/output support
Virtual inputs behave like normal inputs, but their states are determined by the state of multiple inputs rather than being explicitly defined and networked.

This allows for inputs that activate conditionally, such as a braking flash pattern that only activates when Emergency.Warning is on M1-M3, or siren-specific flash patterns that only activate when Emergency.Warning is on M3.