v2.0.3 - Add vehicle custom siren support
Siren slots can be defined using VEHICLE.Siren = { [1] = x }

If x is a string, an existing siren set is used. If x is a table, a new siren set is created with the given tones.

The siren slot can be mapped to a speaker component by setting "Siren = x" in the equipment definition, where x is the siren slot. (Primary siren is 1, secondary is 2, tertiary is 3, etc).

See the Boulder Police FPIU file for an example of implementation.