Add schema querying, HUD and input interfacing
Input schemas are used to define what channels and modes should be displayed to end-users. They are currently defined in a local prototype variable, but will be settable per vehicle (and probably integrated with the library for easier reuse).

User commands and the HUD can "query" the schema to automatically use first entries, last entries, offset entries, cycle, etc. This allows for Photon 2 to adjust input actions and HUD elements based on the schema and supported modes (similar to Photon LE but more versatile).