v2.0.10 Add input sequence ranking
Within a component input mode (like `Emergency.Warning:MODE3`), the `Segment = "SEQUENCE"` syntax can optionally written like `Segment = { "SEQUENCE", Rank = # }`.

The `Rank` parameter specifies the sequence's priority compared to other sequences in MODE3.

This is only necessary if a light is being controlled by two or more sibling sequences in the same mode. If it is not defined and multiple sequences are conflicting, the winning sequence is usually random and will cause unexpected behavior.