Add VEHICLE.Default parameter
Setting `VEHICLE.Default = true` will apply the vehicle profile to all vehicles of the same class and model as the base vehicle defined with `VEHICLE.Vehicle`.

It will also prevent Photon 2 from creating a new (duplicate) vehicle entry for the profile.

This can be used to seamlessly apply Photon 2 vehicle lighting to normal vehicles without any special coding.

(Note that this is experimental and certain things like the HUD and controller beeps are still present.)