Added Fort Collins PD Caprice demo with 2 liveries
Caprice Week continues.

- Added new demonstrator vehicle: the Fort Collins Police Caprice PPV. Modern livery is the default, but the old pre-2013 design is selectable as well.

What makes the new design special here is that the door badge is a hand-made recreation of the actual illustrated badge recreation used on the real thing. It is not a photograph. I must highlight this detail because otherwise absolutely nobody would notice or care about the four hours I spent on it.